[refactor] 서비스 클래스 메소드명 변경에 의한 변경
